requirements.txt
Flask==0.10.1
Flask-WTF==0.11
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.8.1
Werkzeug==0.10.4
WTForms==2.0.2
